% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLPK.R
\docType{data}
\name{ALLPK}
\alias{ALLPK}
\title{A example of 4 PeriodS and 2 Sequences BE data.}
\format{
A data fram with 298 rows and 6 variables.
}
\source{
https://www.ema.europa.eu/en/documents/other/31-annex-ii-statistical-analysis-bioequivalence-study-example-data-set_en.pdf.
}
\usage{
ALLPK
}
\description{
A example of 4 PeriodS and 2 Sequences BE data.
}
\keyword{datasets}
