% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTIDSABE.R
\name{NTIDSABE}
\alias{NTIDSABE}
\title{NTIDSABE
An Average Bioequivalence Procedure For Therapeutic Index Drug}
\usage{
NTIDSABE(Data)
}
\arguments{
\item{Data}{is a data.frame,Data must have four column:Subject,Period,Sequence,PK.The contents of the Sequence should be 1 or 2,1 is Sequence of "TRTR",2 is Sequence of "RTRT".}
}
\value{
resultABE,the result of ABE, PointestABE is GMR,CIABE is 90\% confidence interval of PointestABE.
}
\description{
NTIDSABE
An Average Bioequivalence Procedure For Therapeutic Index Drug
}
\examples{
NTIDSABE(ALLPK)
}
