% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTIDSRSABE.R
\name{NTIDSRSABE}
\alias{NTIDSRSABE}
\title{NTIDSRSABE
An Reference-Scaled Average Bioequivalence Procedure For Therapeutic Index Drug.}
\usage{
NTIDSRSABE(Data)
}
\arguments{
\item{Data}{is a data.frame,Data must have four column:Subject,Period,Sequence,PK.The contents of the Sequence should be 1 or 2,1 is Sequence of "TRTR",2 is Sequence of "RTRT".}
}
\value{
resultRSABE,the result of RSABE, critbound is upper 90\% confidence interval of y.
}
\description{
NTIDSRSABE
An Reference-Scaled Average Bioequivalence Procedure For Therapeutic Index Drug.
}
\examples{
NTIDSRSABE(ALLPK)
}
