% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWR.R
\name{SWR}
\alias{SWR}
\title{SWR
SWR is Evaluate the standard deviation within the individual of the reference Drug.}
\usage{
SWR(Data)
}
\arguments{
\item{Data}{is a data.frame,Data must have four column:Subject,Period,Sequence,PK.The contents of the Sequence should be 1 or 2,1 is Sequence of "TRTR",2 is Sequence of "RTRT".}
}
\value{
ResultSWR is a data.frame, ResultSWR have show the result of S2WR(The square of the standard deviation within the individual),SWR(standard deviation within the individual),DfCVR(The degree of freedom of standard deviation within an individual),CVwR(Intra-individual variation).
}
\description{
SWR
SWR is Evaluate the standard deviation within the individual of the reference Drug.
}
\examples{
SWR(ALLPK)
}
