% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWRATE.R
\name{SWRATE}
\alias{SWRATE}
\title{SWRATE}
\usage{
SWRATE(Data)
}
\arguments{
\item{Data}{is a data.frame,Data must have four column:Subject,Period,Sequence,PK.The contents of the Sequence should be 1 or 2,1 is Sequence of "TRTR",2 is Sequence of "RTRT".}
}
\value{
SWRATE is a data.frame,ResultSWR have show the result of S2WR(The square of the standard deviation within the individual),SWR(standard deviation within the individual),DfCVR(The degree of freedom of standard deviation within an individual),CVwR(Intra-individual variation).
}
\description{
Evaluate the ratio of SWT/SWR and 90\% confidence interval.
}
\examples{
SWRATE(ALLPK)
}
