% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWRATE2.R
\name{SWRATE2}
\alias{SWRATE2}
\title{SWRATE2
use SWR,DFCVR,SWT,DFCVT Evaluate the ratio of SWT/SWR and 90\% confidence interval.}
\usage{
SWRATE2(SWR, DFCVR, SWT, DFCVT)
}
\arguments{
\item{SWR}{The standard deviation within the individual of the reference Drug.}

\item{DFCVR}{The degree of freedom of standard deviation within an individual of the reference Drug.}

\item{SWT}{The standard deviation within the individual of the test Drug.}

\item{DFCVT}{The degree of freedom of standard deviation within an individual of the test Drug.}
}
\value{
SWRATE,sw.ratio is the ratio of SWT/SWR,sw.ratio.CI is the 90\% confidence interval of sw.ratio.
}
\description{
SWRATE2
use SWR,DFCVR,SWT,DFCVT Evaluate the ratio of SWT/SWR and 90\% confidence interval.
}
\examples{
SWRATE2(SWR=0.9,DFCVR=0.75,SWT=0.95,DFCVT=0.67)
}
