% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWT.R
\name{SWT}
\alias{SWT}
\title{SWT.}
\usage{
SWT(Data)
}
\arguments{
\item{Data}{is a data.frame,Data must have four column:Subject,Period,Sequence,PK.The contents of the Sequence should be 1 or 2,1 is Sequence of "TRTR",2 is Sequence of "RTRT".}
}
\value{
ResultSWT is a data.frame, ResultSWT have show the result of S2WT(The square of the standard deviation within the individual),SWT(standard deviation within the individual),DfCVT(The degree of freedom of standard deviation within an individual),CVwT(Intra-individual variation).
}
\description{
SWT is Evaluate the standard deviation within the individual of the test Drug.
}
\examples{
SWT(ALLPK)

}
